Removed old coach registration page
